Moves TransactionHandler to common
