+++ src/common/handlers/ContractEntityHandler.ts
