+++ src/m0_facilitator/Exception.ts
@@ -75,22 +75,6 @@ export class FacilitatorStartException extends Error {
 }
 
 /**
- * This will be thrown when handler implementation not found.
- */
-export class HandlerNotFoundException extends Error {
-  /**
-   * Constructor
-   *
-   * @param message Exception reason.
-   */
-  public constructor(message: string) {
-    super(message);
-    this.name = 'HandlerNotFoundException';
-    this.message = message;
-  }
-}
-
-/**
  * Defines error for facilitator config.
  */
 export class InvalidFacilitatorConfigException extends Error {
