+++ src/m0_facilitator/handlers/stake_and_mint/MintProgressedHandler.ts
@@ -15,12 +15,13 @@
 // ----------------------------------------------------------------------------
 
 
+import ContractEntityHandler from '../../../common/handlers/ContractEntityHandler';
 import Logger from '../../../common/Logger';
-import Message from '../../models/Message';
+
 import {
   MessageDirection, MessageRepository, MessageStatus, MessageType,
 } from '../../repositories/MessageRepository';
-import ContractEntityHandler from '../ContractEntityHandler';
+import Message from '../../models/Message';
 import Utils from '../../Utils';
 
 /**
