+++ test/m0_facilitator/TransactionHandler/handle.test.ts
@@ -17,13 +17,13 @@
 
 import sinon from 'sinon';
 
-import StakeRequestedHandler from '../../../src/m0_facilitator/handlers/stake_and_mint/StakeRequestedHandler';
-import Repositories from '../../../src/m0_facilitator/repositories/Repositories';
-import TransactionHandler from '../../../src/m0_facilitator/TransactionHandler';
-import assert from '../../test_utils/assert';
-import SpyAssert from '../../test_utils/SpyAssert';
-import StubData from '../../test_utils/StubData';
 import { RequestType } from '../../../src/m0_facilitator/repositories/MessageTransferRequestRepository';
+import assert from '../../test_utils/assert';
+import Repositories from '../../../src/m0_facilitator/repositories/Repositories';
+import SpyAssert from '../../test_utils/SpyAssert';
+import StakeRequestedHandler from '../../../src/m0_facilitator/handlers/stake_and_mint/StakeRequestedHandler';
+import StubData from '../../test_utils/StubData';
+import TransactionHandler from '../../../src/common/TransactionHandler';
 
 describe('TransactionHandler.handle()', (): void => {
   const gatewayAddress = '0x0000000000000000000000000000000000000001';
