+++ test/m0_facilitator/subscriptions/Subscriber/subscribe.test.ts
@@ -23,7 +23,7 @@ import GraphClient from '../../../../src/common/subscriptions/GraphClient';
 import SpyAssert from '../../../test_utils/SpyAssert';
 import Subscriber from '../../../../src/common/subscriptions/Subscriber';
 import TransactionFetcher from '../../../../src/common/subscriptions/TransactionFetcher';
-import TransactionHandler from '../../../../src/m0_facilitator/TransactionHandler';
+import TransactionHandler from '../../../../src/common/TransactionHandler';
 
 describe('Subscriber.subscribe()', () => {
   let mockApolloClient: any;
