+++ test/m0_facilitator/subscriptions/Subscriber/unsusbcribe.test.ts
@@ -22,7 +22,7 @@ import ContractEntityRepository from '../../../../src/common/repositories/Contra
 import GraphClient from '../../../../src/common/subscriptions/GraphClient';
 import Subscriber from '../../../../src/common/subscriptions/Subscriber';
 import TransactionFetcher from '../../../../src/common/subscriptions/TransactionFetcher';
-import TransactionHandler from '../../../../src/m0_facilitator/TransactionHandler';
+import TransactionHandler from '../../../../src/common/TransactionHandler';
 
 describe('Subscriber.unsubscribe()', () => {
   let mockApolloClient: any;
